pyproject.toml
[tool.poetry]
name = "interactive"
version = "0.1.0"
description = ""
authors = ["jorgebarcelos <jorgebarcelos.faria@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "4.2.5"
django-extensions = "^3.2.3"
django-ninja = "^0.22.2"
django-cpf-cnpj = "^1.0.0"
django-widget-tweaks = "^1.5.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
